Use stdint

+++ exercises/isogram/src/example.c
@@ -1,18 +1,19 @@
 #include "isogram.h"
 #include <ctype.h>
 #include <stddef.h>
+#include <stdint.h>
 
 bool isIsogram(const char phrase[])
 {
-   long usedLetters = 0;
+   uint32_t usedLetters = 0;
 
    if (phrase == NULL)
       return false;
 
-   for (int i = 0; phrase[i] != '\0'; ++i) {
+   for (int16_t i = 0; phrase[i] != '\0'; ++i) {
       if (isalpha(phrase[i])) {
-         int bitNumber = toupper(phrase[i]) - 'A';
-         long bitMask = 1 << bitNumber;
+         uint8_t bitNumber = toupper(phrase[i]) - 'A';
+         uint32_t bitMask = 1 << bitNumber;
 
          if ((usedLetters & bitMask) > 0)
             return false;
